Deletes all State objects with a name containing the letter 'a'

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -13,7 +13,7 @@ if __name__ == "__main__":
     Session = sessionmaker(bind=engine)
     session = Session()
     del_states = session.query(State).filter(State.name.like("%a")).all()
-    for state in del_states:
-        session.delete(state)
-    session.commit()
-    session.close()
+    if del_states:
+        for state in del_states:
+            session.delete(state)
+        session.commit()
